Encapsulate error resolution update into a lower level function
